Accounts, people and avatars

+++ src/main/java/com/startuppuccino/account/AccountRepository.java
@@ -1,11 +1,12 @@
 package com.startuppuccino.account;
 
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
-
-public interface AccountRepository extends JpaRepository<Account, Long>
+public interface AccountRepository extends JpaRepository<Account, Integer>
 {
     Account findOneByEmail(String email);
+
 }
